+++ src/test/java/org/janelia/saalfeldlab/n5/demo/BlockIterators.java
@@ -81,7 +81,7 @@ public class BlockIterators {
 				.flatMap( shardPosition -> {
 
 					final int nd = attrs.getNumDimensions();
-					final long[] min = attrs.getBlockPositionFromShardPosition(shardPosition, new int[nd]);
+					final long[] min = attrs.getBlockPositionFromShardPosition(shardPosition, new long[nd]);
 					return toStream(new GridIterator(GridIterator.int2long(blocksPerShard), min));
 				});
 	}
